Move to `rate_limiting.hpp`
